+++ app/src/components/MuteParticipants/index.tsx
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
-//
-// SPDX-License-Identifier: EUPL-1.2
-import MuteParticipants from './MuteParticipants';
-
-export default MuteParticipants;